+++ clang/test/dpct/test_api_level/Util/api_test19.cu
@@ -3,13 +3,13 @@
 // RUN: FileCheck --input-file %T/Util/api_test19_out/count.txt --match-full-lines %s
 // RUN: rm -rf %T/Util/api_test19_out
 
-// CHECK: 1
+// CHECK: 2
 // TEST_FEATURE: Util_cmul
 
 #include <cuComplex.h>
 
 __device__ void foo1() {
-    float2 a, b;
+    double2 a, b;
 
     auto c = cuCmul(a, b);
